Move user and agent modules under participant package. Fixes #172
