--- dialoguekit/agent/__init__.py
@@ -1,7 +1,0 @@
-"""Agent level init."""
-from dialoguekit.agent.agent import Agent, AgentType
-
-__all__ = [
-    "Agent",
-    "AgentType",
-]
