+++ dialoguekit/participant/__init__.py
@@ -1,4 +1,9 @@
 """Participant init."""
 from dialoguekit.participant.participant import DialogueParticipant, Participant
+from dialoguekit.participant.user_preferences import UserPreferences
 
-__all__ = ["Participant", "DialogueParticipant"]
+__all__ = [
+    "Participant",
+    "DialogueParticipant",
+    "UserPreferences"
+]
