--- dialoguekit/user/__init__.py
@@ -1,5 +1,0 @@
-"""Module level init for the users."""
-from dialoguekit.user.user import User
-from dialoguekit.user.user_preferences import UserPreferences
-
-__all__ = ["UserPreferences", "User"]
