+++ sample_agents/moviebot_agent.py
@@ -15,10 +15,10 @@ server with the right configuration.
 
 import requests
 
-from dialoguekit.agent.agent import Agent, AgentType
 from dialoguekit.core.annotated_utterance import AnnotatedUtterance
 from dialoguekit.core.intent import Intent
 from dialoguekit.core.utterance import Utterance
+from dialoguekit.participant.agent import Agent, AgentType
 from dialoguekit.participant.participant import DialogueParticipant
 
 _MOVIEBOT_DEFAULT_URI = "http://127.0.0.1:5001"
