+++ sample_agents/rasa_parrot_agent.py
@@ -6,10 +6,10 @@ This agent depends on Rasa parrot project to parrot back. See
 
 import requests
 
-from dialoguekit.agent.agent import Agent
 from dialoguekit.core.annotated_utterance import AnnotatedUtterance
 from dialoguekit.core.intent import Intent
 from dialoguekit.core.utterance import Utterance
+from dialoguekit.participant.agent import Agent
 from dialoguekit.participant.participant import DialogueParticipant
